Remove automatic spell-checking on all files.
While spell-checking is a wonder while writing or editing plain-text, it
is annoying when dealing with code files. Since the majority of files I
deal with on a daily basis are written in Python, HTML, or sh (as opposed
to English), it does not really make sense to have to ":set nospell"
after opening most files.