Symlink script to automate linking dotfiles.
I was originally going to use the Bash script in
<https://github.com/michaeljsmalley/dotfiles> as a template, but I
eventually decided to write a Python script from scratch. This is mainly
because I did not want to license this under the GPL3, but also because
I like Python and felt a bit rusty with using the os and shutil
modules.